Poprawione wykrywanie sesji tmuxa

+++ fish/config.fish
@@ -3,26 +3,29 @@ set -gx EDITOR nvim
 set -g fish_greeting
 
 if status is-interactive
+  echo 1
   if type -q pfetch
     pfetch
   end
 
   if not set -q SSH_CONNECTION;
+    echo 2
     if type -q tmux; and not set -q TMUX; # 1 oznacza że nie znaleziono tej zmiennej
+      echo 3
       set output (tmux list-sessions 2>&1) # Wyświetl listę sesji tmuxa gdy nie jesteś w tmuxie
-      if string match --invert --quiet 'no server running*' "$output"
-        if string match --invert --quiet 'error*' "$output"
+      if string match --invert --quiet 'no server running*' "$output"; and string match --invert --quiet 'error*' "$output"
+          echo 5
           echo Tmux sessions
           echo $output
           set sessions_opened (echo $output | wc -l)
           echo $sessions_opened
           if [ "$sessions_opened" = "1" ]
+            echo 5
             set first_window (echo $output | sed 's/:.*//g')
             echo Only one window so attaching to $first_window
             tmux a -t $first_window
             kill $fish_pid
           end
-        end
       else
         tmux # rozpocznij nową instancję
         kill $fish_pid
